# services/read/email/api/openapi.yaml
openapi: 3.0.3
info:
  title: Cairn Email Ingest Service API
  version: 0.1.0
  description: |
    Email ingestion service for Cairn. Accepts forwarded emails from a
    Cloudflare Email Worker, processes them, and delivers sanitised content
    to the Content Service via an outbox pattern.

    ## Authentication
    - **Ingest endpoint**: API key via `X-API-Key` header
    - **User-facing endpoints**: JWT Bearer token (RS256, key from Vault)

    ## Email Flow
    1. Cloudflare Worker forwards email JSON to `POST /api/v1/source/email/ingest`
    2. Service resolves recipient address to a user
    3. Raw email is stored for background processing
    4. Worker cleans HTML (tracking pixels, footers), sanitises, and delivers to Content Service
  contact:
    name: Cairn API Support
    url: https://github.com/andrew-craig/cairn-reader
  license:
    name: MIT
servers:
  - url: http://localhost:8089
    description: Email Ingest Service (local development, host-mapped port)
...
paths:
...
  /api/v1/source/email/user/{user_id}/senders:
    get:
      tags: [Senders]
      summary: List senders
      description: |
        Returns a paginated list of senders who have emailed the user,
        ordered by most recently received first.
      operationId: listSenders
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserID'
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of senders to return
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of senders to skip
      responses:
        '200':
          description: List of senders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse_SenderList'
        '400':
          description: Invalid user ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: User can only access their own resources
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for the Cloudflare Email Worker
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 JWT token issued by the User Service
  parameters:
    UserID:
      name: user_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: User UUID (must match the authenticated JWT user_id)
  schemas:
...
    SenderResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sender_email:
          type: string
          format: email
        sender_name:
          type: string
          description: Display name (omitted if not known)
        email_count:
          type: integer
          description: Number of emails received from this sender
        last_received_at:
          type: string
          format: date-time
          description: Timestamp of the most recent email (omitted if none processed yet)

    # --- Envelope wrappers ---
    Meta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        version:
          type: string
          example: v1
...
    SuccessResponse_SenderList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/SenderResponse'
        meta:
          $ref: '#/components/schemas/Meta'
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Machine-readable error code
          example: bad_request
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          additionalProperties: true
          description: Optional additional error details
        meta:
          $ref: '#/components/schemas/Meta'